test_dependency_collector: port to generic layer.
