+++ Library/Homebrew/test/test_dependency_collector.rb
@@ -76,16 +76,6 @@ class DependencyCollectorTests < Homebrew::TestCase
     assert_predicate dep, :optional?
   end
 
-  def test_ld64_dep_pre_leopard
-    MacOS.stubs(:version).returns(MacOS::Version.new("10.4"))
-    assert_equal LD64Dependency.new, @d.build(:ld64)
-  end
-
-  def test_ld64_dep_leopard_or_newer
-    MacOS.stubs(:version).returns(MacOS::Version.new("10.5"))
-    assert_nil @d.build(:ld64)
-  end
-
   def test_raises_typeerror_for_unknown_classes
     assert_raises(TypeError) { @d.add(Class.new) }
   end
